Refactor to include predicate folder
