+++ src/main/java/seedu/address/model/predicate/IdContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.filter;
+package seedu.address.model.predicate;
 
 import java.util.List;
