+++ src/main/java/seedu/address/model/predicate/SerializablePredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.filter;
+package seedu.address.model.predicate;
 
 import java.util.function.Predicate;
