+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -17,8 +17,8 @@ import seedu.address.logic.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.filter.SerializablePredicate;
 import seedu.address.model.person.Person;
+import seedu.address.model.predicate.SerializablePredicate;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for
