+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.FindCommand;
-import seedu.address.model.filter.NameContainsKeywordsPredicate;
+import seedu.address.model.predicate.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
